+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - `<ui-menuitem>`: add prop `value`
 - `<ui-form>`: add prop `itemMarginBottom`, `labelWidth`, `labelMarginRight`, `labelMarginBottom`
 - `v-anchor`: optimize for dynamic html
+- `v-shape`: add cut sharps for experimental
 - overwrite variables become more concise and clear
 
   - Old